The HTTP-based backend of the Replication worker services (1)

+++ src/replica/worker/WorkerDeleteReplicaHttpRequest.cc
@@ -40,8 +40,6 @@
 // LSST headers
 #include "lsst/log/Log.h"
 
-#define CONTEXT context("WorkerDeleteReplicaHttpRequest", __func__)
-
 using namespace std;
 namespace fs = std::filesystem;
 using json = nlohmann::json;
@@ -52,36 +50,44 @@ LOG_LOGGER _log = LOG_GET("lsst.qserv.replica.WorkerDeleteReplicaHttpRequest");
 
 }  // namespace
 
+#define _CONTEXT context("WorkerDeleteReplicaHttpRequest::", __func__)
+#define _SET_STATUS_FAILED(extendedStatus_, error_)       \
+    LOGS(_log, LOG_LVL_ERROR, _CONTEXT << " " << error_); \
+    setStatus(lock, protocol::Status::FAILED, extendedStatus_, error_);
+
 namespace lsst::qserv::replica {
 
 shared_ptr<WorkerDeleteReplicaHttpRequest> WorkerDeleteReplicaHttpRequest::create(
         shared_ptr<ServiceProvider> const& serviceProvider, string const& worker,
-        protocol::QueuedRequestHdr const& hdr, json const& req, ExpirationCallbackType const& onExpired) {
+        protocol::QueuedRequestHdr const& hdr, protocol::RequestParams const& params,
+        ExpirationCallbackType const& onExpired) {
     auto ptr = shared_ptr<WorkerDeleteReplicaHttpRequest>(
-            new WorkerDeleteReplicaHttpRequest(serviceProvider, worker, hdr, req, onExpired));
+            new WorkerDeleteReplicaHttpRequest(serviceProvider, worker, hdr, params, onExpired));
     ptr->init();
     return ptr;
 }
 
 WorkerDeleteReplicaHttpRequest::WorkerDeleteReplicaHttpRequest(
         shared_ptr<ServiceProvider> const& serviceProvider, string const& worker,
-        protocol::QueuedRequestHdr const& hdr, json const& req, ExpirationCallbackType const& onExpired)
-        : WorkerHttpRequest(serviceProvider, worker, "DELETE", hdr, req, onExpired),
-          _databaseName(req.at("database")),
-          _chunkNumber(req.at("chunk")),
-          // This status will be returned in all contexts
-          _replicaInfo(ReplicaInfo::Status::NOT_FOUND, worker, _databaseName, _chunkNumber,
-                       util::TimeUtils::now(), ReplicaInfo::FileInfoCollection{}) {}
+        protocol::QueuedRequestHdr const& hdr, protocol::RequestParams const& params,
+        ExpirationCallbackType const& onExpired)
+        : WorkerHttpRequest(serviceProvider, worker, "DELETE", hdr, params, onExpired),
+          _databaseName(params.requiredString("database")),
+          _chunkNumber(params.requiredUInt32("chunk")) {
+    // This status will be returned in all contexts
+    _replicaInfo = ReplicaInfo(ReplicaInfo::Status::NOT_FOUND, worker, _databaseName, _chunkNumber,
+                               util::TimeUtils::now(), ReplicaInfo::FileInfoCollection{});
+}
 
-void WorkerDeleteReplicaHttpRequest::getResult(json& result) const {
-    result["replica_info"] = _replicaInfo.toJson();
+json WorkerDeleteReplicaHttpRequest::getResult() const {
+    return json::object({{"replica_info", _replicaInfo.toJson()}});
 }
 
 bool WorkerDeleteReplicaHttpRequest::execute() {
-    LOGS(_log, LOG_LVL_DEBUG, CONTEXT << " db: " << _databaseName << " chunk: " << _chunkNumber);
+    LOGS(_log, LOG_LVL_DEBUG, _CONTEXT << " db: " << _databaseName << " chunk: " << _chunkNumber);
 
-    replica::Lock lock(_mtx, CONTEXT);
-    checkIfCancelling(lock, CONTEXT);
+    replica::Lock lock(mtx, _CONTEXT);
+    checkIfCancelling(lock, _CONTEXT);
 
     // The method will throw ConfigUnknownDatabase if the database is invalid.
     DatabaseInfo const databaseInfo = serviceProvider()->config()->databaseInfo(_databaseName);
@@ -91,32 +97,33 @@ bool WorkerDeleteReplicaHttpRequest::execute() {
 
     // The data folder will be locked while performing the operation
     int numFilesDeleted = 0;
-    WorkerHttpRequest::ErrorContext errorContext;
     std::error_code ec;
     {
-        replica::Lock dataFolderLock(_mtxDataFolderOperations, CONTEXT);
+        replica::Lock dataFolderLock(mtxDataFolderOperations, _CONTEXT);
         fs::path const dataDir = fs::path(serviceProvider()->config()->get<string>("worker", "data-dir")) /
                                  database::mysql::obj2fs(_databaseName);
         fs::file_status const stat = fs::status(dataDir, ec);
-        errorContext =
-                errorContext ||
-                reportErrorIf(stat.type() == fs::file_type::none, protocol::StatusExt::FOLDER_STAT,
-                              "failed to check the status of directory: " + dataDir.string() +
-                                      ", code: " + to_string(ec.value()) + ", error: " + ec.message()) ||
-                reportErrorIf(!fs::exists(stat), protocol::StatusExt::NO_FOLDER,
-                              "the directory does not exists: " + dataDir.string());
+        if (stat.type() == fs::file_type::none) {
+            _SET_STATUS_FAILED(protocol::StatusExt::FOLDER_STAT,
+                               "failed to check the status of directory: " + dataDir.string() +
+                                       ", code: " + to_string(ec.value()) + ", error: " + ec.message());
+            return true;
+        }
+        if (!fs::exists(stat)) {
+            _SET_STATUS_FAILED(protocol::StatusExt::NO_FOLDER,
+                               "the directory does not exists: " + dataDir.string());
+            return true;
+        }
         for (auto const& name : files) {
             fs::path const file = dataDir / fs::path(name);
             if (fs::remove(file, ec)) ++numFilesDeleted;
-            errorContext = errorContext || reportErrorIf(ec.value() != 0, protocol::StatusExt::FILE_DELETE,
-                                                         "failed to delete file: " + file.string() +
-                                                                 ", code: " + to_string(ec.value()) +
-                                                                 ", error: " + ec.message());
+            if (ec.value() != 0) {
+                _SET_STATUS_FAILED(protocol::StatusExt::FILE_DELETE,
+                                   "failed to delete file: " + file.string() +
+                                           ", code: " + to_string(ec.value()) + ", error: " + ec.message());
+                return true;
+            }
         }
-    }
-    if (errorContext.failed) {
-        setStatus(lock, protocol::Status::FAILED, errorContext.extendedStatus);
-        return true;
     }
     setStatus(lock, protocol::Status::SUCCESS);
     return true;
